+++ mcp_tools/web_app_security/enhanced_web_app_security.py
@@ -4,61 +4,6 @@ from typing import Dict, Any, Optional
 
 def register_enhanced_web_app_security_tools(mcp, hexstrike_client, logger, HexStrikeColors):
 
-
-    @mcp.tool()
-    def anew_data_processing(input_data: str, output_file: str = "",
-                            additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute anew for appending new lines to files (useful for data processing).
-
-        Args:
-            input_data: Input data to process
-            output_file: Output file path
-            additional_args: Additional anew arguments
-
-        Returns:
-            Data processing results with unique line filtering
-        """
-        data = {
-            "input_data": input_data,
-            "output_file": output_file,
-            "additional_args": additional_args
-        }
-        logger.info("📝 Starting anew data processing")
-        result = hexstrike_client.safe_post("api/tools/anew", data)
-        if result.get("success"):
-            logger.info("✅ anew data processing completed")
-        else:
-            logger.error("❌ anew data processing failed")
-        return result
-
-    @mcp.tool()
-    def qsreplace_parameter_replacement(urls: str, replacement: str = "FUZZ",
-                                       additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute qsreplace for query string parameter replacement.
-
-        Args:
-            urls: URLs to process
-            replacement: Replacement string for parameters
-            additional_args: Additional qsreplace arguments
-
-        Returns:
-            Parameter replacement results for fuzzing
-        """
-        data = {
-            "urls": urls,
-            "replacement": replacement,
-            "additional_args": additional_args
-        }
-        logger.info("🔄 Starting qsreplace parameter replacement")
-        result = hexstrike_client.safe_post("api/tools/qsreplace", data)
-        if result.get("success"):
-            logger.info("✅ qsreplace parameter replacement completed")
-        else:
-            logger.error("❌ qsreplace parameter replacement failed")
-        return result
-
     @mcp.tool()
     def uro_url_filtering(urls: str, whitelist: str = "", blacklist: str = "",
                          additional_args: str = "") -> Dict[str, Any]:
